Refactor CLTextInput and ControlledInput components
